getseq/requirements.txt
appnope==0.1.3
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
biopython==1.79
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.2.1
six==1.16.0
stack-data==0.4.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
